Add a multicast TS output option
This is not likely to be generally useful, but I find it useful :D .

The idea is that you can convert a remote TS formatted stream into a
local multicast stream that can be picked up by any number of machines,
so all of you raspberry pis (or whatever) can listen to a multicast
group and watch a single stream.

Currently this uses an installation of PyZeroconf to get the
multicast socket setup. The multicast setup isn't particularly involved,
so it could be recreated easily if someone wanted to avoid the
LGPL module dependency.

cli is the only package that changes. It gains a MulticastOutput
class that just does the packetization and socket setup.